requirements.txt
tensorflow-gpu==1.14.0
requests==2.22.0
Flask==1.1.1
numpy==1.18.1
grpcio==1.27.2
matplotlib==3.1.2
inflect==4.1.0
tensorflow_serving_api==1.14.0
sounddevice==0.3.15
librosa==0.7.2
scipy==1.3.2
tqdm==4.36.1
Unidecode==1.1.1
grpc==1.0.0
hparams==0.3.0
lws==1.2.6
scikit_learn==0.23.1
tensorflow_serving==9.9.0
torch==1.6.0
